Transport changes

+++ simulation/Region.h
@@ -70,7 +70,7 @@ private:
     mutable std::map<Region, double> connections;
     mutable std::map<Region, double> flights;
     std::map<std::string, bool> eventHistory;
-    int **history = nullptr;
+    double **history = nullptr;
     int historyDay = 0;
     int historySize = 0;
     int historyWidth = 0;
@@ -105,13 +105,13 @@ public:
     [[nodiscard]] double getMi() const;
     [[nodiscard]] long int getPopulation() const;
     [[nodiscard]] long int getSusceptible() const;
-    [[nodiscard]] long int getExposed() const;
-    [[nodiscard]] long int getInfectious() const;
+    [[nodiscard]] double getExposed() const;
+    [[nodiscard]] double getInfectious() const;
     [[nodiscard]] long int getRecovered() const;
     [[nodiscard]] long int getDead() const;
     std::map<Region, double>& getConnections() const;
     std::map<Region, double>& getFlights() const;
-    int ** getHistory() const;
+    double ** getHistory() const;
     int getHistorySize() const;
     int getHistoryWidth() const;
     bool getIsHistoryEmpty() const;
